feat: codex plan mode reasoning effort as a managed option
A managed Codex profile can now set `plan_mode_reasoning_effort`, which
governs how hard Codex thinks while it is in plan mode. The profile cannot
put it there: Codex 0.146 has no flag or config key for the startup
collaboration mode, `collaboration_mode` is an app-server per-turn
parameter rather than a config file key, and plan mode is entered with
Shift+Tab or `/plan` in the TUI. The control says exactly that instead of
implying the profile selects the mode. Upstream request: openai/codex#13942.

Adding it pushed managed_agent_launch.rs over the 500-line limit, so the
shell rendering moved to its own module with its own tests. That is a real
boundary: building a launch from a config is one concern, quoting it for
whichever interactive shell the PTY holds is another.
