+++ docs/orchestration.md
@@ -47,6 +47,8 @@ The adapter type is required because the registry is more than a command: it dec
 
 Managed mode exposes the supported model, effort, persona, permission, sandbox, and trust controls for the selected adapter. Missing options deliberately defer to that agent's own configuration. The host validates the structured configuration and quotes the resulting executable and arguments for the actual platform shell only when the terminal is spawned. Command mode remains available for advanced or unsupported CLI options; its value must be the interactive form the adapter expects because a one-shot mode cannot satisfy the accept/heartbeat/complete worker contract. Existing profiles are migrated as Command profiles without changing their behavior.
 
+A managed Codex profile carries a second effort setting, `planModeEffort`, which becomes `-c plan_mode_reasoning_effort=<value>`. It governs only how hard Codex thinks while it is in plan mode, and the profile cannot put it there: Codex 0.146 has no flag or config key for the startup collaboration mode, which is entered with Shift+Tab or `/plan` in the TUI (`collaboration_mode` is an app-server per-turn parameter, not a config file key). The editor says so on the control rather than implying the profile selects the mode. The upstream request is openai/codex#13942.
+
 A managed Claude profile can also pass `--allow-dangerously-skip-permissions` through `allowSkipPermissions`. That flag only makes bypass reachable inside the session; `permissionMode: bypassPermissions` is what starts there, and the two are independent on purpose. It still counts as a reduced-protection marker, so turning it on asks for confirmation before the profile is saved.
 
 A managed Claude profile may also carry an optional `ccsProfile`. When it is set, the launch replaces the executable with `ccs` and passes the profile as its first positional argument, followed by the same Claude flags, because that is the switcher's own contract (`ccs <profile> [claude-args...]`). The value must be a single name that does not start with a dash: it is positional, so a dash-prefixed value would be read as an option of the switcher itself. Only the Claude adapter accepts the key; the host rejects it anywhere else. Note that `ccs` points `CLAUDE_CONFIG_DIR` at its own profile directory, which is where Alera installs the Claude agent-status hooks, so a session launched through a CCS profile reports status only if that profile inherits those hooks.
