+++ rust/alera-cli/src/terminal_host/orchestration/managed_agent_launch.rs
@@ -16,6 +16,9 @@ pub struct ManagedAgentLaunch {
 /// change.
 const CCS_EXECUTABLE: &str = "ccs";
 
+/// One reasoning-effort enum serves both of Codex's effort settings.
+const CODEX_EFFORTS: &[&str] = &["minimal", "low", "medium", "high", "xhigh", "max", "ultra"];
+
 pub fn build_managed_agent_launch(
     agent_type: &str,
     config: &Value,
@@ -48,29 +51,13 @@ pub fn build_managed_agent_launch(
     })
 }
 
-pub fn render_managed_launch(launch: &ManagedAgentLaunch, shell: &str) -> String {
-    let family = shell_family(shell);
-    std::iter::once(launch.executable.as_str())
-        .chain(launch.arguments.iter().map(String::as_str))
-        .map(|value| quote_argument(value, family))
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-pub fn managed_launch_preview(launch: &ManagedAgentLaunch) -> String {
-    #[cfg(windows)]
-    let shell = std::env::var("COMSPEC").unwrap_or_else(|_| "cmd.exe".to_string());
-    #[cfg(not(windows))]
-    let shell = std::env::var("SHELL").unwrap_or_else(|_| "/bin/sh".to_string());
-    render_managed_launch(launch, &shell)
-}
-
 fn build_codex(values: &Map<String, Value>, arguments: &mut Vec<String>) -> Result<(), String> {
     require_known_keys(
         values,
         &[
             "model",
             "effort",
+            "planModeEffort",
             "sandbox",
             "approvalPolicy",
             "webSearch",
@@ -78,14 +65,18 @@ fn build_codex(values: &Map<String, Value>, arguments: &mut Vec<String>) -> Resu
         ],
     )?;
     push_string(values, "model", "--model", arguments)?;
-    if let Some(effort) = enum_value(
-        values,
-        "effort",
-        &["minimal", "low", "medium", "high", "xhigh", "max", "ultra"],
-    )? {
+    if let Some(effort) = enum_value(values, "effort", CODEX_EFFORTS)? {
         arguments.extend([
             "--config".to_string(),
             format!("model_reasoning_effort={effort}"),
+        ]);
+    }
+    // Codex has no launch-time way to enter plan mode, so this only sets how
+    // hard it thinks once the user is in it.
+    if let Some(effort) = enum_value(values, "planModeEffort", CODEX_EFFORTS)? {
+        arguments.extend([
+            "--config".to_string(),
+            format!("plan_mode_reasoning_effort={effort}"),
         ]);
     }
     let bypass = bool_value(values, "bypassApprovalsAndSandbox")?;
@@ -459,36 +450,6 @@ fn push_non_negative_integer(
         .ok_or_else(|| format!("{key} must be a non-negative integer."))?;
     arguments.extend([flag.to_string(), number.to_string()]);
     Ok(())
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ShellFamily {
-    Posix,
-    PowerShell,
-    Cmd,
-}
-
-fn shell_family(shell: &str) -> ShellFamily {
-    let normalized = shell.replace('\\', "/").to_ascii_lowercase();
-    let executable = normalized.rsplit('/').next().unwrap_or(&normalized);
-    if matches!(
-        executable,
-        "powershell" | "powershell.exe" | "pwsh" | "pwsh.exe"
-    ) {
-        ShellFamily::PowerShell
-    } else if matches!(executable, "cmd" | "cmd.exe") {
-        ShellFamily::Cmd
-    } else {
-        ShellFamily::Posix
-    }
-}
-
-fn quote_argument(value: &str, family: ShellFamily) -> String {
-    match family {
-        ShellFamily::Posix => format!("'{}'", value.replace('\'', "'\"'\"'")),
-        ShellFamily::PowerShell => format!("'{}'", value.replace('\'', "''")),
-        ShellFamily::Cmd => format!("\"{}\"", value.replace('"', "\"\"")),
-    }
 }
 
 #[cfg(test)]
