+++ rust/alera-cli/src/terminal_host/orchestration/mod.rs
@@ -7,6 +7,7 @@ pub mod dispatch_preamble;
 pub mod group_resolution;
 pub mod lifecycle_reconciliation;
 pub mod managed_agent_launch;
+pub mod managed_launch_shell_rendering;
 pub mod message_delivery;
 pub mod message_formatter;
 pub mod message_waiters;
